Add packaging logic for Fedora-family of OSes
The goal of the added script is to generate a source package, in this
case a source RPM that is 'fully contained' in a sense that it does
not depend on arbitrary network access for building, but only
on the specified spec dependencies that are provided by the build
environment and the rest of the pre-built packages in the Linux
distribution, in this case - Fedora.

Under a git clone of this package, simply run:

    ./packaging/build-srpm -o .

It will generate a Source RPM of HEAD that can be built locally
using rpmbuild or mock, or under Fedora's build servers.

For example, an build-srpm run has generated:

    haxe-3.2.0-2015.07.23.234315.git1a7d770290ec.fc21.src.rpm

Subsequently, running the Fedora 'mock' tool on the source RPM
has generated:

    haxe-3.2.0-2015.07.23.234315.git1a7d770290ec.fc21.src.rpm
    haxe-3.2.0-2015.07.23.234315.git1a7d770290ec.fc21.x86_64.rpm
    haxe-debuginfo-3.2.0-2015.07.23.234315.git1a7d770290ec.fc21.x86_64.rpm

Tested with Fedora 21.

This install paths of the haxe RPM are /usr/bin, and /usr/share/haxe.